renamed Algorithm trait to Filter
