+++ src/tests/test_kalman_filter.rs
@@ -1,6 +1,6 @@
 use crate::state::State;
 use crate::measurement::Measurement;
-use crate::common::Algorithm;
+use crate::filter::Filter;
 use crate::algorithms::KalmanFilter;
 use crate::common::na as na;
 use crate::common::Epoch;
@@ -9,7 +9,7 @@ use crate::common::Epoch;
 fn test_kalman_update_scalar() {
     // Small 1D Kalman filter: N=1, M=1
     // bring Algorithm trait into scope so we can call the trait-associated `new`
-    let mut filter = <KalmanFilter::<1, 1> as Algorithm>::new();
+    let mut filter = <KalmanFilter::<1, 1> as Filter>::new();
 
     // initial state x = 0
     assert_eq!(filter.x.value[(0, 0)], 0.0);
@@ -38,7 +38,7 @@ fn test_kalman_update_scalar() {
 #[test]
 fn test_kalman_update_2d() {
     // N=2, M=1. Observe only the first state element.
-    let mut filter = <KalmanFilter::<2, 1> as Algorithm>::new();
+    let mut filter = <KalmanFilter::<2, 1> as Filter>::new();
     // set H = [1 0]
     filter.h = na::SMatrix::<f64, 1, 2>::from_row_slice(&[1.0, 0.0]);
 
@@ -59,7 +59,7 @@ fn test_kalman_update_2d() {
 #[should_panic(expected = "innovation covariance S is singular")]
 fn test_kalman_singular_innovation() {
     // Construct filter where p and r are zero so S == 0 and inverse fails.
-    let mut filter = <KalmanFilter::<1, 1> as Algorithm>::new();
+    let mut filter = <KalmanFilter::<1, 1> as Filter>::new();
     filter.p = na::SMatrix::<f64, 1, 1>::zeros();
     filter.r = na::SMatrix::<f64, 1, 1>::zeros();
     filter.h = na::SMatrix::<f64, 1, 1>::identity();
